ViewController reused for DiningList,CollectionList and PublicArtsList and Resolved collectionlist bug
